pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "probable_fiesta"
authors = [{ name = "Sergio Munoz", email = "sergio.munoz@pubnub.com" }]
description = "Python Core Package."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.7"
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
]


dependencies = [
  "python-dotenv",
  "click>=8.0.3",
]
dynamic = ["version"]

[project.urls]
"Homepage" = "https://github.com/sergio-munoz/probable_fiesta/"
"Bug Tracker" = "https://github.com/sergio-munoz/probable_fiesta/issues"
Documentation = "https://github.com/sergio-munoz/probable_fiesta#readme"
Issues = "https://github.com/sergio-munoz/probable_fiesta/issues"
Source = "https://github.com/sergio-munoz/probable_fiesta"

[project.scripts]
probable-fiesta = "probable_fiesta.run:main"

[tool.hatch.version]
path = "src/probable_fiesta/__about__.py"


[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]",
  #"pytest",
  #"pytest-cov",
]

#[tool.hatch.build]
#include = [
    #"probable_fiesta/*.py",
    #"tests/",
#]
#[tool.hatch.build.targets.wheel.hooks.mypyc]
#dependencies = ["hatch-mypyc"]

#[build.targets.wheel.hooks.mypyc]
#include = ["/src/probable_fiesta"]
#exclude = ["__main__.py"]

[tool.hatch.envs.default.scripts]
#cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=pubnub_pytools --cov=tests"
#no-cov = "cov --no-cov"
cov = "coverage run --omit tests/*,scripts/*  -m unittest"
report = "coverage report"
probable-fiesta = "probable_fiesta.run:main"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "src/probable_fiesta/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
